virtual-friend-bot/requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.8.2
uvloop==0.20.0; platform_system != "Windows"
